Improve style.
